Deploy backend, add email auth, permissions onboarding, and simulator toggle
- Deploy Cloudflare Worker backend to ski-tracker-api.apexdiligence.workers.dev
- Create D1 database with migrations, set JWT secret and Google OAuth client ID
- Add email/password registration and login (PBKDF2 hashing) alongside Google OAuth
- Add permission onboarding flow (location, background GPS, notifications, battery, motion)
- Add foreground service for GPS tracking with screen off
- Add simulator toggle banner on session screen for emulator testing
- Change airtimeMs from double to int for precise millisecond tracking
- Fix login screen to pop on auth success and on skip
- Fix API client to handle non-JSON error responses gracefully
- Remove unused environment_sensors dependency that blocked Android build

+++ lib/features/session/providers/session_providers.dart
@@ -198,7 +198,7 @@ class SessionNotifier extends StateNotifier<SessionState> {
     if (_currentSessionId != null) {
       final jumps = _recorder.jumps;
       final maxAirtime =
-          jumps.isEmpty ? 0.0 : jumps.map((j) => j.airtimeMs).reduce(max);
+          jumps.isEmpty ? 0.0 : jumps.map((j) => j.airtimeMs.toDouble()).reduce(max);
 
       await _sessionRepo.finishSession(
         id: _currentSessionId!,
